// src/routes/_authenticated/_shell/business/navigation/app-sidebar-type.ts
import type { LucideIcon } from "lucide-react";

export type NavId =
  | "workspace"
  | "comic-playground"
  | "member-list"
  | "system-mail"
  | "utilities"
  | "settings";

export type NavConfig = {
  id: NavId;
  path: string;
  label: string;
  icon: LucideIcon;
};

export type TeamConfig = {
  id: string;
  name: string;
  short: string;
  desc: string;
  avatarUrl: string;
  avatarThumbnailUrl?: string | undefined;
};